chore: fixing sonar issues KER-346
